Douban Top250 books info extracted by scrapy
This is a simple scrapy crawling the topic250 book info from douban, the start_url is https://book.douban.com/top250?start=0。
And the info is stored into database tables created by Django Admin Site. You can find the code in my repository: 'the django book'-->mysite. In the mysite project, there is an app named douban which is the for the scrapy project. It is very simple and a good example for new beginners.